Refactor ModalChecklistDetails component and styles
- Consolidated and updated the ModalChecklistDetails component to improve readability and maintainability.
- Removed unused imports and code related to checklist deletion.
- Enhanced file upload handling and checklist item management.
- Updated styles for better layout and responsiveness.
- Removed the deprecated ModalChecklistDetails2 component and its styles.
